Refactor email funcionality.
