+++ src/AppBundle/Controller/ProjectController.php
@@ -240,115 +240,6 @@ class ProjectController extends Controller
           'edit_form' => $editForm->createView(), ]);
     }
 
-    /*---------------------------------------RETRIEVE EMAIL FUNTIONALLITY----------------------------------*/
-    /**
-     * Displays a form to testing send email with TEST email.
-     *
-     * @Route("/retrieve-email", name="retrieve_email")
-     * @Method("GET")
-     */
-    public function retrieveEmailAction()
-    {
-        $imapFlags = '{imap.gmail.com:993/imap/ssl/novalidate-cert}INBOX';
-
-        $emailUser = 'i.delgado@paralleldevs.com';
-
-        $passEmail = 'IADMa1992';
-
-        $em = $this->getDoctrine()->getManager();
-
-        $mailbox = imap_open($imapFlags, $emailUser, $passEmail) or die(imap_last_error());
-
-        $numMessages = imap_num_msg($mailbox);
-
-        for ($i = 1; $i <= $numMessages; ++$i) {
-            $headers = imap_headerinfo($mailbox, $i);
-
-            if ($headers->Unseen === 'U') {
-                $fromAddr = $headers->from[0]->mailbox.'@'.$headers->from[0]->host;
-
-                $project_type = $em->getRepository('AppBundle:ProjectType')->findBy(['name' => 'Support']);
-
-                $ticketTypeArray = $em->getRepository('AppBundle:TicketType')->findBy(['name' => 'Created by email']);
-
-                $ticketStatusArray = $em->getRepository('AppBundle:TicketStatus')->findBy(['name' => 'Open']);
-
-                $projects = $em->getRepository('AppBundle:Project')
-                    ->findBy(['email' => $fromAddr,
-                        'projectType' => $project_type,
-                    ]);
-
-                $imap_uid = imap_uid($mailbox, $i);
-
-                $ticket_DB = $em->getRepository('AppBundle:Ticket')->findByIdEmailTicket($imap_uid);
-
-                if (sizeof($ticket_DB) == 0) {
-                    $ticket = new Ticket();
-                    $ticket->setName($headers->subject);
-                    $body_msg = imap_fetchbody($mailbox, $i, 1);
-                    $ticket->setDescription($body_msg);
-                    $ticket->setCreatedAt(new \DateTime('now'));
-                    $ticket->setIdEmailTicket($imap_uid);
-
-                    foreach ($ticketTypeArray as $ticketType) {
-                        $ticket->setTicketType($ticketType);
-                    }
-                    foreach ($ticketStatusArray as $ticketStatus) {
-                        $ticket->setTicketStatus($ticketStatus);
-                    }
-                    foreach ($projects as $project) {
-                        $ticket->setProject($project);
-                    }
-
-                    $em->persist($ticket);
-                    $em->flush();
-                }
-            }
-        }
-
-        $all_tiquets = $em->getRepository('AppBundle:Ticket')->findAll();
-
-        return $this->render('Project/formSendEmailTestDos.html.twig', ['entity' => $all_tiquets]);
-    }
-
-    /**
-     * Delete all elements in the table.
-     *
-     * @Route("/delete-all", name="delete_all_elements")
-     * @Method("GET")
-     */
-    public function deleteAllTicketsAction()
-    {
-        $em = $this->getDoctrine()->getManager();
-
-        $entity = $em->getRepository('AppBundle:Ticket')->findAll();
-
-        foreach ($entity as $ticket) {
-            $em->remove($ticket);
-        }
-
-        $em->flush();
-
-        if (!$entity) {
-            throw $this->createNotFoundException('All elements was deleted  ');
-
-            $em->remove($entity);
-            $em->flush();
-        }
-
-        return $this->render('Project/formSendEmailTestDos.html.twig', ['entity' => $entity]);
-    }
-        /*---------------------------------------RETRIEVE EMAIL FUNTIONALLITY----------------------------------*/
-    /**
-     * Displays a form to create a new Project entity.
-     *
-     * @Route("/form-email", name="new_form_email")
-     * @Method("GET")
-     */
-    public function newFormEmailAction()
-    {
-        return $this->render('Project/formSendEmailTest.html.twig');
-    }
     /**
      * Deletes a Project entity.
      *
